move `compute lint` and `compute lintstats` to their own directories.
