+++ cmd/registry/cmd/compute/compute.go
@@ -17,6 +17,8 @@ package compute
 import (
 	"github.com/apigee/registry/cmd/registry/cmd/compute/complexity"
 	"github.com/apigee/registry/cmd/registry/cmd/compute/conformance"
+	"github.com/apigee/registry/cmd/registry/cmd/compute/lint"
+	"github.com/apigee/registry/cmd/registry/cmd/compute/lintstats"
 	"github.com/spf13/cobra"
 )
 
@@ -28,8 +30,8 @@ func Command() *cobra.Command {
 
 	cmd.AddCommand(conformance.Command())
 	cmd.AddCommand(complexity.Command())
-	cmd.AddCommand(lintCommand())
-	cmd.AddCommand(lintStatsCommand())
+	cmd.AddCommand(lint.Command())
+	cmd.AddCommand(lintstats.Command())
 	cmd.AddCommand(scoreCommand())
 	cmd.AddCommand(scoreCardCommand())
 	cmd.AddCommand(vocabularyCommand())
