+++ cmd/registry/cmd/compute/lint/lint-openapi-spectral.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package compute
+package lint
 
 import (
 	"encoding/json"
